moved spectrum to extras module
