+++ code/ulab.h
@@ -34,6 +34,6 @@
 #define ULAB_FILTER_MODULE (1)
 
 // user-defined modules
-#define ULAB_EXTRAS_MODULE (0)
+#define ULAB_EXTRAS_MODULE (1)
 
 #endif
